En principio model, repository, service y controller está. He cambiado las clases primitivas de los atributos a clases encapsuladoras para que no den error al ser null. Por último, he empezado con el tema de Login y Register, he creado AuthService y AuthController para llevar la lógica pero cambiar la contraseña creo que irá en UsuarioService. También he creado el paquete config para tener las configuraciones. Por ahora sólo está SegConfig que lleva el tema de la encriptación en un Bean para que el ciclo de vida lo controle spring.

+++ ángela_fernández_gonzález_proyecto_final/src/main/java/model/Categoria.java
@@ -12,7 +12,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int idCategoria;
+    private Long idCategoria;
 
     @NotNull
     @NotEmpty
@@ -35,7 +35,7 @@ public class Categoria {
         this.imagen = imagen;
     }
 
-    public int getIdCategoria() {
+    public Long getIdCategoria() {
         return idCategoria;
     }
